requirements.txt
langchain
langchain-openai
python-dotenv
scikit-learn
numpy 